+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### Grow
 
+- Set arity on combined pred function based on argument pred functions. Allows checking arity before creating/passing `getFrom` function.
+
 ### Fix
+
+- Pass `options` to `key` pred.
 
 ### Deprecate